Add types file for ContractContext
